Add files via upload
This data science project was done on census data of a moderately sized town sandwiched between two cities sourced online. The census data was being analysed to find out what should be built on an unoccupied plot of land which the local government wishes to develop and what the government should invest in. The data provided was explored, cleaned and formatted appropriately. Thereafter charts were plotted for visualization of the clean data to get the true position of the town in certain areas using certain criteria. Based on observations from the charts plotted, recommendations were made to guide the government appropriately.